backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pymongo==4.6.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
wechatpy==2.0.0
python-dotenv==1.0.1
cryptography==42.0.5
requests==2.31.0 